refactor: simplify import path
