+++ src/api/interface/population.ts
@@ -1,5 +1,5 @@
-import { PopulationChange } from "../../domain/polulationChange";
-import { Prefecture } from "../../domain/prefecture";
+import { PopulationChange } from "~/domain/polulationChange";
+import { Prefecture } from "~/domain/prefecture";
 
 export interface PopulationChangeAPI {
   fetchPopulationChange(prefs: Array<Prefecture>): Promise<PopulationChange>;
