+++ src/api/interface/pref.ts
@@ -1,4 +1,4 @@
-import { Prefecture } from "../../domain/prefecture";
+import { Prefecture } from "~/domain/prefecture";
 
 export interface PrefectureAPI {
   fetchPrefectures(): Promise<Array<Prefecture>>;
