+++ src/pages/__test__/index.spec.tsx
@@ -1,5 +1,5 @@
 import renderer from "react-test-renderer";
-import Home from "..";
+import Home from "~/pages/index";
 
 describe("index page", () => {
   test("renders correctly", () => {
